// src/_services/api/model/itensContabilDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API de Gestão de Devolução e Produtividade
 * Documentação da API responsável por: gestão de produtividade, montagem e impressão de mapas de separação, sistema de devolução de estoque e contagem de inventário.
 * OpenAPI spec version: 1.0
 */
import type { ItensContabilDtoTipo } from './itensContabilDtoTipo';
import type { ItensContabilDtoTipoDevolucao } from './itensContabilDtoTipoDevolucao';

export interface ItensContabilDto {
  demandaId: number;
  sku: string;
  descricao: string;
  tipo: ItensContabilDtoTipo;
  tipoDevolucao: ItensContabilDtoTipoDevolucao;
  quantidadeCaixas: number;
  quantidadeUnidades: number;
  avariaCaixas: number;
  avariaUnidades: number;
}